git-svn-id: https://teamsupport.svn.beanstalkapp.com/main/trunk@1558 02885c34-e171-44ab-8435-42fb7a3c3aaf

+++ TeamSupport.Data/BusinessObjects/RecentlyViewedItems.cs
@@ -32,8 +32,7 @@ namespace TeamSupport.Data
           using (SqlCommand command = new SqlCommand())
           {
               command.CommandText =
-                @"DELETE FROM RecentlyViewedItems
-                WHERE (refID = @orgID) AND (refType = 1)";
+                @"DELETE FROM RecentlyViewedItems WHERE (refID = @orgID) AND (refType = 1)";
               command.CommandType = CommandType.Text;
               command.Parameters.AddWithValue("@orgID", organizationID);
               command.ExecuteNonQuery();
